准备immortwrt-v8配置

+++ immortalwrt-diy-init.sh
@@ -1,8 +1,8 @@
 #!/bin/bash
 
-# Set default theme to luci-theme-argon
-uci set luci.main.mediaurlbase='/luci-static/argon'
-uci set luci.sauth.sessiontime=36001
+# 主题相关设置
+[ -d /usr/lib/lua/luci/view/themes/argon ] && uci set luci.main.mediaurlbase='/luci-static/argon'
+uci set luci.sauth.sessiontime=86401
 uci commit luci
 
 # Disable IPV6 ula prefix
@@ -18,15 +18,29 @@ mkdir -p /_LXusb/5T
 grep -qw LXusb /etc/passwd || echo "LXusb:*:1000:1000:USB User:/_LXusb:/bin/false" >> /etc/passwd
 grep -qw LXusb /etc/group || echo "LXusb:x:1000:USB User" >> /etc/group
 
-rm /usr/bin/wget
-ln -s /usr/libexec/wget-ssl /usr/bin/wget
+[ -f /etc/init.d/collectd ] &&  mkdir -p /etc/collectd/conf.d/
 
-/etc/init.d/nlbwmon enable
+if [ -f /usr/libexec/wget-ssl ]; then
+	rm /usr/bin/wget
+	ln -s /usr/libexec/wget-ssl /usr/bin/wget
+fi
 
-cp /etc/config/wrtbwmon.user /etc/
+[ -f /etc/init.d/nlbwmon ] && /etc/init.d/nlbwmon enable
+[ -f /etc/init.d/blockd ] && /etc/init.d/blockd disable
+[ -f /etc/init.d/netdata ] && /etc/init.d/netdata disable
+[ -f /etc/init.d/openclash ] && /etc/init.d/openclash disable
+[ -f /etc/init.d/passwall_server ] && /etc/init.d/passwall_server disable
+[ -f /etc/init.d/passwall ] && /etc/init.d/passwall disable
 
-mv /usr/bin/AdGuardHome /usr/bin/AdGuardHome_t
-mkdir /usr/bin/AdGuardHome/
-mv /usr/bin/AdGuardHome_t /usr/bin/AdGuardHome/AdGuardHome
+[ -f /etc/config/wrtbwmon.user ] && cp /etc/config/wrtbwmon.user /etc/
+
+if [ -f /usr/bin/AdGuardHome ]; then
+	mv /usr/bin/AdGuardHome /usr/bin/AdGuardHome_t
+	mkdir /usr/bin/AdGuardHome/
+	mv /usr/bin/AdGuardHome_t /usr/bin/AdGuardHome/AdGuardHome
+fi
+if [ -f /usr/bin/AdGuardHome_new ]; then
+	mv /usr/bin/AdGuardHome_new /usr/bin/AdGuardHome/AdGuardHome
+fi
 
 exit 0
